+++ graphql-gateway/src/resolvers/users/user.mutation.js
@@ -1,44 +1,80 @@
+import bcrypt from 'bcryptjs'
+
 import { isString, isNumber } from 'lodash'
 
+import authUtils from '../../utils/auth'
+import passwordUtils from '../../utils/password'
+
 const UserMutation = {
-  async createUser(parent, { data }, { userService, logger }) {
-    logger.info('UserMutation#createUser.call', data)
+  async signup(parent, { data }, { userService, logger }) {
+    logger.info('UserMutation#signup.call', data)
 
     const userExists = (await userService.count({ where: { email: data.email } })) >= 1
 
-    logger.info('UserMutation#createUser.check', userExists)
+    logger.info('UserMutation#signup.check', userExists)
 
     if (userExists) {
       throw new Error('Email taken')
     }
 
-    const user = await userService.create(data)
+    const password = await passwordUtils.hashPassword(data.password)
 
-    logger.info('UserMutation#createUser.result', user)
+    const user = await userService.create({
+      ...data,
+      password
+    })
 
-    return user
+    delete user.password
+
+    logger.info('UserMutation#signup.result', user)
+
+    return {
+      user,
+      token: authUtils.generateToken(user.id)
+    }
   },
-  async updateUser(parent, args, { userService, logger }) {
-    const { id, data } = args
+  async login(parent, { data }, { userService, logger }) {
+    logger.info('UserQuery#login.call', data)
 
-    logger.info('UserMutation#updateUser.call', id, data)
+    const user = await userService.findOne({
+      where: {
+        email: data.email
+      }
+    })
 
-    const user = await userService.findOne({ where: { id } })
-
-    logger.info('UserMutation#updateUser.target', user)
+    logger.info('UserQuery#login.check1', !user)
 
     if (!user) {
-      throw new Error('User not found')
+      throw new Error('Unable to login')
     }
 
-    if (isString(data.email)) {
-      const userExists = (await userService.count({ where: { email: data.email } })) >= 1
+    const isMatch = await bcrypt.compare(data.password, user.password)
 
-      if (userExists) {
-        throw new Error('Email taken')
-      }
+    logger.info('UserQuery#login.check2', !user)
 
-      user.email = data.email
+    if (!isMatch) {
+      throw new Error('Unable to login')
+    }
+
+    delete user.password
+
+    logger.info('UserQuery#login.result', user)
+
+    return {
+      user,
+      token: authUtils.generateToken(user.id)
+    }
+  },
+  async updateProfile(parent, { data }, { request, userService, logger }) {
+    logger.info('UserMutation#updateProfile.call', data)
+
+    const id = await authUtils.getUser(request)
+    const user = await userService.findOne({ where: { id } })
+
+    logger.info('UserMutation#updateProfile.target', user)
+
+    if (!user) {
+      throw new Error('User profile not found')
     }
 
     if (isString(data.name)) {
@@ -51,22 +87,104 @@ const UserMutation = {
 
     const updatedUser = await userService.update(id, user)
 
-    logger.info('UserMutation#updateUser.result', updatedUser)
+    delete updatedUser.password
+
+    logger.info('UserMutation#updateProfile.result', updatedUser)
 
     return updatedUser
   },
-  async deleteUser(parent, { id }, { userService, logger }) {
-    logger.info('UserMutation#deleteUser.result', id)
+  async updateEmail(parent, { data }, { request, userService, logger }) {
+    logger.info('UserMutation#updateEmail.call', data)
 
+    const id = await authUtils.getUser(request)
     const user = await userService.findOne({ where: { id } })
+    const isMatch = await bcrypt.compare(data.currentPassword, user.password)
+
+    logger.info('UserMutation#updateEmail.target', user)
+    logger.info('UserMutation#updateEmail.check1', !user || !isMatch)
+
+    if (!user || !isMatch) {
+      throw new Error('Error updating email. Kindly check the email or password provided')
+    }
+
+    const userExists = (await userService.count({ where: { email: data.email } })) >= 1
+
+    logger.info('UserMutation#updateEmail.check2', userExists)
+
+    if (userExists) {
+      throw new Error('Email taken')
+    }
+
+    user.email = data.email
+
+    const updatedUser = await userService.update(id, user)
+
+    delete updatedUser.password
+
+    logger.info('UserMutation#updateEmail.result', updatedUser)
+
+    return {
+      updatedUser,
+      token: authUtils.generateToken(user.id)
+    }
+  },
+  async updatePassword(parent, { data }, { request, userService, logger }) {
+    logger.info('UserMutation#updatePassword.call', data)
+
+    const id = await authUtils.getUser(request)
+    const user = await userService.findOne({ where: { id } })
+    const isMatch = await bcrypt.compare(data.currentPassword, user.password)
+    const isConfirmed = data.newPassword === data.confirmPassword
+
+    logger.info('UserMutation#updatePassword.target', user)
+    logger.info('UserMutation#updatePassword.check', !user || !isMatch || !isConfirmed)
+
+    if (!user || !isMatch || !isConfirmed) {
+      throw new Error('Error updating password. Kindly check your passwords.')
+    }
+
+    const password = await passwordUtils.hashPassword(data.password)
+
+    const updatedUser = await userService.update(id, {
+      ...user,
+      password
+    })
+
+    delete updatedUser.password
+
+    logger.info('UserMutation#updatePassword.result', updatedUser)
+
+    return {
+      updatedUser,
+      token: authUtils.generateToken(user.id)
+    }
+  },
+  async deleteAccount(parent, args, { request, commentService, postService, userService, logger }) {
+    logger.info('UserMutation#deleteAccount.call')
+
+    const id = await authUtils.getUser(request)
+    const user = await userService.findOne({ where: { id } })
+
+    logger.info('UserMutation#deleteAccount.check1', !user)
 
     if (!user) {
       throw new Error('User not found')
     }
 
+    const postExists = (await postService.count({ where: { author: id } })) >= 1
+    const commentExists = (await commentService.count({ where: { author: id } })) >= 1
+
+    logger.info('UserMutation#deleteAccount.check1', postExists || commentExists)
+
+    if (postExists || commentExists) {
+      throw new Error('You have already made posts and comments. Kindly delete those first.')
+    }
+
     const count = await userService.destroy(id)
 
-    logger.info('UserMutation#deleteUser.result', count, user)
+    delete user.password
+
+    logger.info('UserMutation#deleteAccount.result', count, user)
 
     return count
   }
